::Boggle:: The RStarTree has been storing positions as integers all along. This would certainly make it impossible to match the findNeighbors metric. Patched to use doubles.
Former-commit-id: 2a450c7c8f76a1eb30b872bd2daa2d5506a26f26

+++ ClusterFinder/RStarTree/RStarBoundingBox.h
@@ -30,15 +30,15 @@ template <std::size_t dimensions>
 struct RStarBoundingBox {
 
 	// edges[x].first is low value, edges[x].second is high value
-	std::pair<int, int> edges[dimensions];
+	std::pair<double, double> edges[dimensions];
 	
 	// forces all edges to their extremes so we can stretch() it
 	void reset()
 	{
 		for (std::size_t axis = 0; axis < dimensions; axis++)
 		{
-			edges[axis].first = std::numeric_limits<int>::max();
-			edges[axis].second = std::numeric_limits<int>::min();
+			edges[axis].first = std::numeric_limits<double>::max();
+			edges[axis].second = std::numeric_limits<double>::min();
 		}
 	}
 	
@@ -78,7 +78,7 @@ struct RStarBoundingBox {
 	// the sum of all deltas between edges
 	inline int edgeDeltas() const
 	{
-		int distance = 0;
+		double distance = 0;
 		for (std::size_t axis = 0; axis < dimensions; axis++)
 			distance += edges[axis].second - edges[axis].first;
 			
@@ -127,10 +127,10 @@ struct RStarBoundingBox {
 		for (std::size_t axis = 0; area && axis < dimensions; axis++)
 		{
 			// this makes it easier to understand
-			const int x1 = edges[axis].first;
-			const int x2 = edges[axis].second;
-			const int y1 = bb.edges[axis].first;
-			const int y2 = bb.edges[axis].second;
+			const double x1 = edges[axis].first;
+			const double x2 = edges[axis].second;
+			const double y1 = bb.edges[axis].first;
+			const double y2 = bb.edges[axis].second;
 		
 			// left edge outside left edge
 			if (x1 < y1)
